Move test eip2930 to test_sign
